test(e2e): split node-killing HA test into its own composed suite file (#24503)
Round-2 e2e consolidation, PR 9/9 — owns `composed/**`.

## Clock-skew suite stays on the composed HA Postgres cluster
The four `Clock Skew and Timezone Safety` its in
`composed/ha/e2e_ha_full.parallel.test.ts` exercise
`PostgresSlashingProtectionDatabase` directly against the running 5-node
HA cluster's **real dockerized PostgreSQL** slashing-protection DB:

- TZ-independent duty timestamps (absolute-time storage)
- `cleanupOldDuties` keeps recent duties when the node clock is 2h ahead
- `cleanupOldDuties` deletes old duties by DB time even when the node
clock is 1h behind
- `cleanupOwnStuckDuties` keeps recent stuck duties when the node clock
is 3h ahead

The property under test is that these DB-clock semantics
(`CURRENT_TIMESTAMP`, `timestamptz`) are immune to node clock/timezone
skew — the node and its database must be able to genuinely diverge in
clock and timezone environment for the assertion to mean anything. That
only holds against a real, separate Postgres process. An earlier
revision moved these into an in-process PGlite file; since PGlite runs
inside the Node process, the DB and node can never actually diverge, so
that variant is dropped. The tests keep riding the composed cluster's
docker Postgres via the shared `HaFullTestContext` (`t.mainPool` /
`t.dateProvider`), exactly as before; `dateProvider` simulates the
skewed node clock while the DB uses its own clock. The
`@electric-sql/pglite` dependency added for the PGlite variant is
removed from end-to-end (other workspaces that use pglite for unit tests
are unaffected).

## Order-dependent HA test → own file
The `should distribute work across multiple HA nodes` test was annotated
`must run last` because it permanently kills every node. Extracted
verbatim into `composed/ha/e2e_ha_distribute_work.test.ts`, which shares
the cluster setup with the remaining suite via a new non-test
`ha_full_setup.ts` module (the ~380-line `beforeAll`/`afterAll`/helpers
moved verbatim into an `HaFullTestContext` class). Result:

- The killer test gets its own cluster; the ordering contract is gone
(no hidden reliance on definition order).
- It has a single `it`, so it is a plain `.test.ts` (not `.parallel`)
and runs as one whole-file CI container.
- The remaining `e2e_ha_full.parallel.test.ts` keeps its 3
non-destructive its (block production, governance voting, keystore
reload) plus the clock-skew describe, all order-independent.
- Zero CI runtime change: setup was moved, not rewritten, so behavior is
unchanged. No assertions dropped.

Added a `.test_patterns.yml` flaky entry for the new file mirroring the
existing `e2e_ha_full` one (owner unchanged).

## Stranded composed tests — left in place, still excluded
`composed/e2e_persistence.test.ts` and
`composed/integration_proof_verification.test.ts` are excluded from
every CI list and have run nowhere since April 2025. I moved each into a
category dir, ran it, and reverted the move because both are broken on
the current branch — root causes outside this diff's editable scope:

| File | Disposition | Why |
|---|---|---|
| `integration_proof_verification.test.ts` | left in `composed/`,
excluded, documented | Committed
`fixtures/dumps/epoch_proof_result.json` is stale (last regenerated Feb
2026; circuits/VK changed since). bb and the on-chain HonkVerifier both
reject the proof (`Failed to verify RootRollupArtifact proof!`). Needs
the fixture regenerated; better relocated to the bb-prover circuit
tests. |
| `e2e_persistence.test.ts` | left in `composed/`, excluded, documented
| `beforeAll` no longer completes: the single-node sequencer stalls in
checkpoint proposal (`waitForAttestationsAndEnqueueSubmissionAsync`) and
the 600s hook times out. The root cause is in shared setup/sequencer,
not this file. |

Each file's header comment and the `bootstrap.sh` exclusion now document
the real reason (previously "excluded for unknown reasons"). **Decision
items for follow-up:** regenerate the epoch-proof fixture and relocate
`integration_proof_verification` to bb-prover; fix the single-node
checkpoint stall and refile `e2e_persistence` under `single-node/`.

## Verified locally
- `yarn build` (full), `yarn format end-to-end`, `yarn lint end-to-end`
— all clean.
- Confirmed both stranded tests fail on this branch (evidence above),
which is why they stay excluded.
- The compose/ha suites need docker (Postgres/Web3Signer) and were not
run locally; the `e2e_ha_full` / `e2e_ha_distribute_work` split and the
clock-skew describe rely on CI. Setup is moved verbatim and each
`.parallel` it is already isolated per container, so runtime behavior is
unchanged.
